missing column added

+++ etl/enh/enh_610_pipeline.py
@@ -131,7 +131,7 @@ class ENHPipeline(EasyPipeline):
             "how_much_think_cost_from_third_home_annualized":     "UInt32",
             "how_much_think_cost_donated_annualized":             "UInt32",
             "how_much_think_cost_other_annualized":               "UInt32",
-            "did_home_get_product_service":                       "UInt32",
+            "did_home_get_product_service":                       "UInt8",
             "get_product_service_buy":                            "UInt8",
             "get_product_service_self_consumption":               "UInt8",
             "get_product_service_self_supply":                    "UInt8",
